Commit 3 - Main Menu Navigation Complete
Completed the Tweens and UI navigation within the main menu making use of my tween scripts Unity event calls.
Created a UI Manager to add additional functionality and set up a swtich state machine to cover problems with tween script also to be used later in UI State Actions.

Next steps are to create a colour picker for the effect settings and begin development on the object script and object data.