Commit 4 - Colour Picker, Player Controls and Game Manager
Created the Colour Picker for the effect settings.
Added Running and interaction controls to the player and input scripts.
Begain the development of the interaction between the Player, Input Manager, Game Manager, UI Manager and UI State Actions Scritps to allow for the proper play loop. accessing the game and UI alongside designing for the future implimentation of other states like Terraforming and Camera Rail.

Next Steps, Research variable and data manipulation as I am struggling to get the required information into the game manager as it is not longer inherited to the monobehavour class.
keep to make sure state are correct, asset menu and asset menu settings UI states might need to become game states as they are accessed directly from game and not the main menu UI.
Continue adding the other player controls.
Continue designing and developing the Object script with Sound, Model and Effect data and functions keeping in mind serialization, saving, loading and databasing.