Commit 7 - Beggings of Asset Settings
Started designing the asset settings interaction.

Next steps - Make the Object script contain a function that opens and closes the asset settings and performs all other relivant tasks. This function is this assigned to the interaction delegate on the player to be called upon using the interaction key. A raycast will be used to assign the function to the delegate from the centre of the screen. A reticle might be useful