Commit 9.1 - Changed method for environment and post processing settings selection.
Previously the buttons where assigned to one of 4 different functions. with the addition of the post processing setting the size of the script would of become cumbersome if the methodology continued. I am now using a string reference switch statement to increase or decrease relivant date to manage setting selection.

Next Steps: Impliment the selecting of post processing data changing and displaying relivant information and impliment the struct and functionality relivant to playing settings including saving them via player prefs.