Commit 12.1 - Fixes Audio mute settings
Made the audio mute setting disable the interaction of the volume sliders so yuor unable to set volumes when its meant to be muted.
Adjusted a problem with the state machine resulting in null references for the player and input manager within the created state actions on scene creation

Next step: Solve null reference of the Ui Manager begin main menu tween when called from Inputmanager. I propose it is due to the Input manager being the caller in a difference scene to the tween script resulting in a null reference even tho call is done via the game manager - state action - Ui manager serialize variable