Commit 15.1 - Accessbility
I removed the UI elements for Colour Blind mode for 2 reasons. Firstly I am unable to truely fine any guidance on implimenting such feature instead finding filters to let see your game through the eyes of something with certain types of colour blindness to design around them. This doesn't work for my application as the designing is in the hands of the user, I need to ensure that my UI in friendly to all users. The second reason is that my application doesn't hold any competion against players or even AI so the requirement to differ between colours is null. instead its to enjoy being in an environment.