Commit 16.9 - Object Move and Rotatiin
Current Bug is that the movement is very jagged but lerping the 2 vector3 didn't achieve anything. I need to take this up with someone more experience to figure out another methodlogy or problem with my code.
I have removed the requirement of scaling the objects as this would effect the titling of the texture casuing distorion on models.