Commit 20.3 - Spawning Object redesign
Redesigned the spawning of objects to allow for moving, position, rotating and spawing multiple objects easier.

Next Steps: Add the ability to delete objects easily and add the ability to shuffle through the list of models under the same item.

Bugs: The raycast that is ment to be pointing from center screen doesn't place the object at centre screen.
The object that is selected from the asset menu for some reason doesn't allow placment of copied right away and requires a few inputs before noticing. It doesnt make sense since the movement of the objects in the same function works but the recognisation of an input on the mouse isn't. Might be a problem with the input manager after coming out of the asset menu but the scene data base script comes out of the asset menu from the game state machine so inputs are changes and everything SHOULD work but it doesn't